Middleware for filtering cards, checking permissions, refactoring, frontend webcam, posibility to add image of card

+++ api/routes/web.php
@@ -17,11 +17,15 @@ $router->get('/', function () use ($router) {
 
 //Business cards
 $router->get('/b-cards', 'BusinessCardController@getAll');
-$router->get('/b-cards/user/{user_id}', 'BusinessCardController@getByUser');
+$router->get('/b-cards/user/{user_id}', ['middleware'=> 'business_cards_filter', 'uses' => 'BusinessCardController@getByUser']  );
 $router->get('/b-cards/{id}', 'BusinessCardController@getById');
 $router->post('/b-cards/', 'BusinessCardController@add');
-$router->put('/b-cards/{id}', 'BusinessCardController@update');
-$router->delete('/b-cards/{id}', 'BusinessCardController@delete');
+$router->put('/b-cards/{id}', ['uses'=>'BusinessCardController@update', 'middleware' => 'check_permission']);
+$router->delete('/b-cards/{id}', ['uses'=>'BusinessCardController@delete', 'middleware' => 'check_permission']);
+
+//Users
+$router->get('/users/', 'UserController@getAll');
+$router->get('/users/{id}', 'UserController@getById');
 
 //Permissions
 $router->get('/permissions', 'PermissionController@getAll');
